refactor(manifest): split helm.go into per-kind files
Iter-16 senior-eng audit secondary: pkg/manifest/helm.go was the only
file in pkg/manifest over ~240 LOC (444 LOC), jamming three unrelated
Flux CRs together. Every other kind in the package already has its
own file (bucket.go, configmap.go, kustomization.go, resourceset.go,
secret.go, source.go); HelmRelease + HelmChartSource + HelmRepository
were the outliers.

Split into:

  helmrelease.go     (343 LOC) HelmRelease, HelmChart inline projection,
                               HelmChartLookup, ResolveChartRef,
                               ParseHelmRelease, chartFromHelmRelease,
                               helmChartFromSource, shortenReleaseName,
                               GetLabels / GetAnnotations accessors.
  helmchartsource.go  (59 LOC) HelmChartSource CR + ParseHelmChartSource.
  helmrepository.go   (49 LOC) HelmRepository CR + ParseHelmRepository.

Each file's imports are scoped to what it actually uses (helmrelease
drops sourcev1; helmchartsource keeps it for HelmChartSpec;
helmrepository keeps it for HelmRepositorySpec).

No behavior change. Package surface unchanged.
